Consolidate the damage section onto the CommonSections
